+++ allauth/socialaccount/signals.py
@@ -1,24 +1,32 @@
 from django.dispatch import Signal
 
 
-# Sent after a user successfully authenticates via a social provider,
+# Emitted after a user successfully authenticates via a social provider,
 # but before the login is actually processed. This signal is emitted
 # for social logins, signups and when connecting additional social
 # accounts to an account.
-# Provides the arguments "request", "sociallogin"
+# Arguments:
+# - request: HttpRequest
+# - sociallogin: SocialLogin
 pre_social_login = Signal()
 
-# Sent after a user connects a social account to a their local account.
-# Provides the arguments "request", "sociallogin"
+# Emitted after a user connects a social account to their local account.
+# Arguments:
+# - request: HttpRequest
+# - sociallogin: SocialLogin
 social_account_added = Signal()
 
-# Sent after a user connects an already existing social account to a
+# Emitted after a user connects an already existing social account to
 # their local account. The social account will have an updated token and
 # refreshed extra_data.
-# Provides the arguments "request", "sociallogin"
+# Arguments:
+# - request: HttpRequest
+# - sociallogin: SocialLogin
 social_account_updated = Signal()
 
-# Sent after a user disconnects a social account from their local
+# Emitted after a user disconnects a social account from their local
 # account.
-# Provides the arguments "request", "socialaccount"
+# Arguments:
+# - request: HttpRequest
+# - socialaccount: SocialAccount
 social_account_removed = Signal()